reduce number of functions on OpenTelemetryInstance
