+++ examples/telescope-app/app/src/main/java/io/embrace/opentelemetry/kotlin/telescope/telemetry/AppTracerProvider.kt
@@ -2,8 +2,7 @@ package io.embrace.opentelemetry.kotlin.telescope.telemetry
 
 import android.content.res.Resources
 import io.embrace.opentelemetry.kotlin.ExperimentalApi
-import io.embrace.opentelemetry.kotlin.OpenTelemetryInstance
-import io.embrace.opentelemetry.kotlin.decorateJavaApi
+import io.embrace.opentelemetry.kotlin.toOtelKotlinApi
 import io.embrace.opentelemetry.kotlin.tracing.Tracer
 import io.opentelemetry.exporter.logging.LoggingSpanExporter
 import io.opentelemetry.sdk.OpenTelemetrySdk
@@ -36,7 +35,7 @@ class AppTracerProvider(
             .setTracerProvider(sdkTracerProvider)
             .build()
 
-        val kotlinApi = OpenTelemetryInstance.decorateJavaApi(javaSdk)
+        val kotlinApi = javaSdk.toOtelKotlinApi()
 
         return kotlinApi.tracerProvider.getTracer("AppTracer")
     }
