+++ opentelemetry-kotlin-compat/api/opentelemetry-kotlin-compat.api
@@ -1,13 +1,14 @@
+public final class io/embrace/opentelemetry/kotlin/OpenTelemetryCompatExtKt {
+	public static final fun toOtelJavaApi (Lio/embrace/opentelemetry/kotlin/OpenTelemetry;)Lio/opentelemetry/api/OpenTelemetry;
+}
+
 public final class io/embrace/opentelemetry/kotlin/OpenTelemetryInstanceCompatKt {
 	public static final fun createOpenTelemetryKotlin (Lio/embrace/opentelemetry/kotlin/OpenTelemetryInstance;Lkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function1;Lio/embrace/opentelemetry/kotlin/Clock;Lio/embrace/opentelemetry/kotlin/factory/SdkFactory;)Lio/embrace/opentelemetry/kotlin/OpenTelemetry;
 	public static synthetic fun createOpenTelemetryKotlin$default (Lio/embrace/opentelemetry/kotlin/OpenTelemetryInstance;Lkotlin/jvm/functions/Function1;Lkotlin/jvm/functions/Function1;Lio/embrace/opentelemetry/kotlin/Clock;Lio/embrace/opentelemetry/kotlin/factory/SdkFactory;ILjava/lang/Object;)Lio/embrace/opentelemetry/kotlin/OpenTelemetry;
-	public static final fun decorateJavaApi (Lio/embrace/opentelemetry/kotlin/OpenTelemetryInstance;Lio/opentelemetry/api/OpenTelemetry;Lio/embrace/opentelemetry/kotlin/factory/SdkFactory;)Lio/embrace/opentelemetry/kotlin/OpenTelemetry;
-	public static synthetic fun decorateJavaApi$default (Lio/embrace/opentelemetry/kotlin/OpenTelemetryInstance;Lio/opentelemetry/api/OpenTelemetry;Lio/embrace/opentelemetry/kotlin/factory/SdkFactory;ILjava/lang/Object;)Lio/embrace/opentelemetry/kotlin/OpenTelemetry;
-	public static final fun decorateKotlinApi (Lio/embrace/opentelemetry/kotlin/OpenTelemetryInstance;Lio/embrace/opentelemetry/kotlin/OpenTelemetry;)Lio/opentelemetry/api/OpenTelemetry;
 }
 
-public final class io/embrace/opentelemetry/kotlin/OpenTelemetryJavaExtKt {
-	public static final fun toOtelJavaApi (Lio/embrace/opentelemetry/kotlin/OpenTelemetry;)Lio/opentelemetry/api/OpenTelemetry;
+public final class io/embrace/opentelemetry/kotlin/OtelJavaOpenTelemetryExtKt {
+	public static final fun toOtelKotlinApi (Lio/opentelemetry/api/OpenTelemetry;)Lio/embrace/opentelemetry/kotlin/OpenTelemetry;
 }
 
 public final class io/embrace/opentelemetry/kotlin/context/ContextExtKt {
