+++ opentelemetry-kotlin-compat/src/jvmMain/kotlin/io/embrace/opentelemetry/kotlin/OpenTelemetryCompatExt.kt
@@ -9,7 +9,7 @@ import io.embrace.opentelemetry.kotlin.tracing.OtelJavaTracerProviderAdapter
  * to the opentelemetry-java API.
  *
  * End-users should generally not use this function and should call [createOpenTelemetryKotlin]
- * or [decorateJavaApi] instead.
+ * or [toOtelKotlinApi] instead.
  */
 @ExperimentalApi
 public fun OpenTelemetry.toOtelJavaApi(): OtelJavaOpenTelemetry {
