+++ opentelemetry-kotlin-compat/src/jvmTest/kotlin/io/embrace/opentelemetry/kotlin/framework/OtelKotlinHarness.kt
@@ -5,10 +5,10 @@ import io.embrace.opentelemetry.kotlin.OpenTelemetry
 import io.embrace.opentelemetry.kotlin.OpenTelemetryInstance
 import io.embrace.opentelemetry.kotlin.aliases.OtelJavaIdGenerator
 import io.embrace.opentelemetry.kotlin.createOpenTelemetryKotlin
-import io.embrace.opentelemetry.kotlin.decorateKotlinApi
 import io.embrace.opentelemetry.kotlin.factory.CompatSdkFactory
 import io.embrace.opentelemetry.kotlin.factory.CompatTracingIdFactory
 import io.embrace.opentelemetry.kotlin.factory.TracingIdFactory
+import io.embrace.opentelemetry.kotlin.toOtelJavaApi
 import kotlin.random.Random
 
 @OptIn(ExperimentalApi::class)
@@ -24,7 +24,7 @@ internal class OtelKotlinHarness : OtelKotlinTestRule() {
     }
 
     val javaApi by lazy {
-        OpenTelemetryInstance.decorateKotlinApi(kotlinApi)
+        kotlinApi.toOtelJavaApi()
     }
 }
